+++ cucumber-expressions/CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+* [JavaScript] Added `ExpressionFactory`, which is now the preferred way to create `Expression` instances.
+  In case an expression references undefined parameter types, an `UndefinedParameterTypeExpression` instance
+  will be returned, which matches nothing (its `#match` method always returns null).
+
 ### Changed
 
 ### Deprecated
+
+* [JavaScript] `new CucumberExpression()` and `new RegularExpression()`. Use `ExpressionFactory#createExpression` instead.
 
 ### Removed